Added proper support for the Wither & fixed bugs
A corrupted Wither will now regenerate health similar to how a 'vanilla'
Wither does it.
Made the health a corrupted Wither regenerates per second configurable.
Removed the AttackRadius option in the LightningAura ability, using the
generic MaximumRange would be best.
Updated the CheckEntityHealth task, since we now use setMaxHealth and
actually let the boss health reflect on the health of the entity.