Vastly improved the /cor stats command
Credits: bm01 did all the work on this and approved us using it.
The handling of the players.dat file now happens on a seperate thread,
so that using /cor stats when your file is big won't cause lag anymore.
All the bosses will now also be removed when the plugin disables.