Fixed ArmorPierce & more
Fixed the players.dat having multiple lines for every player (and adding
1 more each time the file is saved). Now the file will just replace the
line if it already exists.
ArmorPierce now gets the EntityDamageEvent from the listener itself
again instead of relying on livingEntity.getLastDamageCause().
Improved the stats command when getting the stats of an individual
player.
The indidividual list will now show the boss names the same way as in
the viewer. Bosses with a different internal name, but the same external
name are combined when shown.
Withers can no longer be spawned using the /cor spawn command, since
this results in unexpected behaviour. In the /cor spawn list, all
withers are striked through and a message informs someone trying to
spawn a wither that it is not available.