Cleanups, improvements and a bugfix
Cleaned up the code in the CorConfigUpdater class to be more efficient
and if a new version is released that changes nothing in the config
files, the ConfigUpdater doesn't have to be updated.

Added the option to disable the automatic reloading of the server after
installing a new update. If configured this way, the user of the command
will just get a message stating that either a reload or a restart is
required for the changes to take effect.

Fixed the bug where Armor Pierce would give an error when trying to use
it on a LivingEntity other than a Player. It can't be used on anything
other than a Player, so we check for that.

Updated the version number to 2.1.2