Use the custom name for livingEntities and more
Added the option to use the custom name feature in minecraft 1.5
(includes the ability to use the health/maxhealth and a healthbar in the
name).
Added the OnDeath activation condition for some abilities, if an ability
doesn't support it and someones specifies it, Corruption will notify the
console.
Added the Command ability, now you can make a Corrupted issue a command.
The help command now only shows the commands for which you have the
required permission.
Added the /cor version command.
Added the respective cor.version permission.
Added a method in the API to spawn a boss/multiple bosses and one to get
all the available bossnames grouped by EntityType.