Improved /cor spawn command
It now allows Console/Command Block usage using a set of coordinates or a Player (uses the Location the Player is looking at).
Also added the possibility for a player to spawn a boss using coordinates.
Last but not least, ok, the look on my face determined that was a lie. Minor command usage derp fixed.