Added Lightning effect option to Summon (default: false)
Added default Summon parameters.
Also added a check if MonsterType was set for the Summon ability and if not, don't consider it a valid ability and warn the user.

Fixed Snare's BlockBreakEvent listener and clone behaviour compatibility. Now, not the initial ability, but the one assigned to a specific Boss registers it's own listener and deregisters it in it's finalize method (called by the garbage collector).